1-jobs_bot :poop: reg resume

+++ repository/storage.go
@@ -5,6 +5,6 @@ import "github.com/MorZLE/jobs_bot/model"
 type Storage interface {
 	Set(student model.Student) error
 	Get(id int64) (model.Student, error)
-	Delete(id int) error
+	Delete(id int64) error
 	Close()
 }
